service-workers/service-worker/postmessage-to-unloaded-client.https.html

<!doctype html>
<title>Service Worker: postMessage to unloaded Client</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
promise_test(t => {
  const url = 'resources/blank.html';
  const script = 'resources/postmessage-to-unloaded-client-worker.js';
  const scope = 'resources/';
  let registration;
  let unload_client;
  return service_worker_unregister_and_register(t, script, scope)
    .then(r => {
        t.add_cleanup(() => { r.unregister(); });
        registration = r;
        return wait_for_state(t, r.installing, 'activated');
      })
    .then(() => {
        return with_iframe(url);
      })
    .then(f => {
        unload_client = f;
        f.contentWindow.navigator.serviceWorker.onmessage = t.step_func(() => {
          throw new Error('The client should have been unloaded.');
        });
        const promise =
            new Promise(resolve => navigator.serviceWorker.onmessage = resolve);
        registration.active.postMessage('syn');
        return promise;
      })
    .then(e => {
        if (e.data == 'ack') {
          // Unload the client the service worker will post a message to.
          unload_client.remove();
          const promise = new Promise(
              resolve => navigator.serviceWorker.onmessage = resolve);
          registration.active.postMessage('post-back-to-unloaded-client');
          t.step_timeout(() => {
            throw new Error('postMessage to unloaded Client did not throw.');
          }, 2000);
          return promise;
        }
      })
    .then(e => {
        assert_equals(
          e.data,
          'InvalidStateError',
          'postMessage to unloaded Client must throw an "InvalidStateError" ' +
          'DOMException.');
      })
    .catch(unreached_rejection(t));
}, 'postMessage from service worker to unloaded Client.');
</script>
</body>
